+++ frontend/src/pages/Record/RecordCreate/RecordCreatePage.tsx
@@ -1,4 +1,4 @@
-import { RecordCreateForm } from "./RecordCreateForm";
+import { RecordCreateForm } from "../../../components/record/recordCreate/RecordCreateForm";
 import { useState } from 'react';
 import { Sport } from "types/calendar";
 import { useNavigate } from "react-router-dom";
